+++ src/Translator.php
@@ -31,7 +31,8 @@ final class Translator implements TranslatorInterface
 
     /**
      * @param string $locale Default locale to use if locale is not specified explicitly.
-     * @param string|null $fallbackLocale Locale to use if message for the locale specified was not found. Null for none.
+     * @param string|null $fallbackLocale Locale to use if message for the locale specified was not found. Null for
+     * none.
      * @param EventDispatcherInterface|null $eventDispatcher Event dispatcher for translation events. Null for none.
      */
     public function __construct(
@@ -86,7 +87,7 @@ final class Translator implements TranslatorInterface
     /**
      * @psalm-immutable
      */
-    public function withCategory(string $category): self
+    public function withDefaultCategory(string $category): self
     {
         if (!isset($this->categorySources[$category])) {
             throw new RuntimeException('Category with name "' . $category . '" does not exist.');